+++ PSTeachingTools.psm1
@@ -6,4 +6,4 @@
 #create a global variable with PLU data
 $vegetableplu = Import-CSV $PSScriptRoot\plu.csv
 
-Export-ModuleMember -Variable vegetableplu
+Export-ModuleMember -Variable vegetableplu -Alias *